requirements.txt
Flask==1.1.1
numpy==1.18.1
Pillow==7.2.0
tensorflow==2.3.0
